refactor() : exception 관련 패키지 global 패키지로 이동
