+++ src/main/java/gift/global/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.global.exception;
 
 import org.springframework.http.HttpStatus;
