+++ src/main/java/gift/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.global.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
