+++ src/test/java/gift/ProductControllerApiTest.java
@@ -3,7 +3,7 @@ package gift;
 import gift.persistence.entity.Product;
 import gift.presentation.dto.RequestProductDto;
 import gift.presentation.dto.ResponseProductDto;
-import gift.exception.NotFoundException;
+import gift.global.exception.NotFoundException;
 import gift.persistence.repository.ProductRepository;
 import java.util.Objects;
 import org.junit.jupiter.api.Test;
